refactor: rename afishe to afishe-page
