+++ src/components/afishe-page/regular-day/afishe-regular-day.tsx
