+++ src/pages/afishe/index.tsx
@@ -1,13 +1,13 @@
 import {NextPage} from 'next';
 import cn from 'classnames/bind';
 
-import {AfisheTitle} from 'components/afishe/title';
-import {FestivalDays} from 'components/afishe/festival-days';
-import {RegularDays} from 'components/afishe/regular-days';
+import {AfisheTitle} from 'components/afishe-page/title';
+import {FestivalDays} from 'components/afishe-page/festival-days';
+import {RegularDays} from 'components/afishe-page/regular-days';
 import {AppLayout} from 'components/app-layout';
 import data from './assets/mock-data.json';
 
-import styles from 'components/afishe/afishe.module.css';
+import styles from 'components/afishe-page/afishe.module.css';
 
 const cx = cn.bind(styles);
